"check_style_score -> 0 of 1 passed. style score: 0/1 "

+++ m4/Assignment-3/Solution.java
@@ -30,7 +30,7 @@ public class Solution
         	n = n/10;
         	count1 += 1;
         }
-        String sum1 = Double.toString(sum);
+        String sum1 = Double.toString((int)sum);
         return sum1;
 	}
